♻️ Refactor: move all character actions into a folder
