+++ client/src/app/store/ApplicationAction.ts
@@ -1,4 +1,4 @@
 import { NightWatchConfigAction } from '../../features/config/NightWatchConfigAction';
-import { CharacterAction } from '../../features/character/CharacterAction';
+import { CharacterAction } from '../../features/character/actions/CharacterAction';
 
 export type ApplicationAction = NightWatchConfigAction | CharacterAction;
